+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fixed multisite action buttons not working
 - Removed outdated PHP 5.4 warning
 - Added `read_timeout` support to Credis
+- Display connection parameters when using Credis
 
 ## 2.0.4